# src/app/api/model/currentUserView.ts
/**
 * MEANS
 * Api dla MEANS
 *
 * OpenAPI spec version: 1.0.0
 * Contact: jarek.frydrych@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface CurrentUserView {
    username?: string;
    firstname?: string;
    surname?: string;
}