# swagger/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: MEANS
  description: Api dla MEANS
  contact:
    email: jarek.frydrych@gmail.com
basePath: /api
tags:
  - name: Security
    description: Endpointy dla modułu Security
  - name: Core
    description: Endpointy dla modułu Core
schemes:
  - http
  - https
paths:
  /security/login:
    x-swagger-router-controller: security
    x-swagger-router-no-firewall: true
    post:
      summary: Logowanie do aplikacji
      operationId: loginHandler
      tags:
        - Security
      parameters:
        - name: content
          description: Dane logowania użytkownika
          in: body
          required: true
          schema:
            $ref: '#/definitions/LoginParameters'
      responses:
        200:
          description: Zalogowano poprwanie
        403:
          description: Niepoprawne dane logowania
  /security/logout:
    x-swagger-router-controller: security
    get:
      summary: Wylogowanie z aplikacji
      operationId: logoutHandler
      tags:
        - Security
      responses:
        200:
          description: Poprawnie wylogowano użytkownika
  /security/current:
    x-swagger-router-controller: security
    get:
      summary: Pobieranie danych zalogowanego użytkownika
      operationId: currentUserQuery
      tags:
        - Security
      responses:
        200:
          description: Dane zalogowanego użytkownika
          schema:
            $ref: '#/definitions/CurrentUserView'
  /security/user:
    x-swagger-router-controller: security
    post:
      summary: Tworzenie usera
      operationId: userNewHandler
      tags:
        - Security
      parameters:
        - name: content
          description: Dane usera
          in: body
          required: true
          schema:
            $ref: '#/definitions/UserNew'
      responses:
        200:
          description: Utworzono użytkownika
  /asd:
    x-swagger-router-controller: core
    get:
      summary: Asd
      operationId: asdHandler
      tags:
        - Core
      responses:
        200:
          description: Asd response
definitions:
  LoginParameters:
    title: LoginParameters
    properties:
      username:
        title: Nazwa użytkownika
        type: string
      password:
        title: Hasło użytkownika
        type: string
  UserNew:
    title: UserNew
    properties:
      username:
        title: Nazwa użytkownika
        type: string
      password:
        title: Hasło użytkownika
        type: string
  CurrentUserView:
    title: CurrentUserView
    properties:
      username:
        title: Nazwa użytkownika
        type: string
      firstname:
        title: Imię użytkownika
        type: string
      surname:
        title: Nazwisko użytkownika
        type: string